PTFS PT 5913982 reconcile holds upon item delete. Nothing had to be done for messaging, just a syspref turned on. This also incidentally fixes occasion where patron work libraries won't save outside of step 3; that needed to be fixed to save the patron's messaging preferences, in this case, CANCELLED_HOLDS

+++ C4/Members.pm
@@ -785,16 +785,17 @@ sub ModMember {
         if C4::Context->preference("BorrowersLog");
 
    # a staff member's work libraries
-   $sth = $dbh->prepare("DELETE FROM borrower_worklibrary
-   WHERE borrowernumber = ?");
-   $sth->execute($data{borrowernumber});
-   foreach(@$worklibraries) {
-       $sth = $dbh->prepare("INSERT INTO borrower_worklibrary
-       (borrowernumber,branchcode) VALUES (?,?)");
-       $sth->execute($data{borrowernumber},$_);
+   if (defined $worklibraries) {
+      $sth = $dbh->prepare("DELETE FROM borrower_worklibrary
+      WHERE borrowernumber = ?");
+      $sth->execute($data{borrowernumber});
+      foreach(@$worklibraries) {
+          $sth = $dbh->prepare("INSERT INTO borrower_worklibrary
+          (borrowernumber,branchcode) VALUES (?,?)");
+          $sth->execute($data{borrowernumber},$_);
+      }
    }
-
-    return $execute_success;
+   return $execute_success;
 }
 
 sub GetWorkLibraries
